Refactor repositories package name
